Add docker-compose.yml for easy testing
Requiring the developer to have a database in order to run tests raises
a barrier to entry in contributing to this repository. We can simplify
it by adding a `docker-compose.yml` file to fire up a
properly-configured PostgreSQL database that is used by default. Also
add a quick blurb about how to use it and run tests in the README.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>