validator: prepend setter instead of overwriting
In order to catch JSON parsing errors when setting validated attributes,
the validator patches in a setter. However, it does this via
`class_eval`, thereby overwriting any setter that already exists at that
level of the class hierarchy. Use prepend instead to be less
destructive, and allow existing setters to continue running.

Fix #66

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>