Use keyword argument for serialization coder
Positional argument is deprecated and logs warnings:

```
DEPRECATION WARNING: Passing the coder as positional argument is deprecated and will be removed in Rails 7.2.

Please pass the coder as a keyword argument:

  serialize :data, coder: JSON
 (called from block (4 levels) in <top (required)> at /Users/dharsha/repos/activerecord_json_validator/spec/json_validator_spec.rb:44)
DEPRECATION WARNING: Passing the coder as positional argument is deprecated and will be removed in Rails 7.2.

Please pass the coder as a keyword argument:

  serialize :other_data, coder: JSON
 (called from block (4 levels) in <top (required)> at /Users/dharsha/repos/activerecord_json_validator/spec/json_validator_spec.rb:45)
```