Use json_schemer 2.0
This upgrades to json_schemer 2.0, which is the latest major version. It
includes support for all current JSON Schema versions and new keywords
(eg, `unevaluatedProperties`). There are also new features that users of
this library may find helpful:

- Symbol key support for schemas and data
- Custom error messages with i18n or `x-error` keyword (it would be
  great to get feedback on this from your users)
- Global configuration options

Using 2.0 will likely require a major version bump because it comes with
breaking changes. See the [changelog][0] for more details.

Closes:
- https://github.com/mirego/activerecord_json_validator/issues/72
- https://github.com/mirego/activerecord_json_validator/issues/73

[0]: https://github.com/davishmcclurg/json_schemer/blob/main/CHANGELOG.md